Add files via upload
This repository contains the source code for a responsive admission form built with HTML, CSS, and JavaScript. The form is designed to collect student information, including personal details, educational background, and more. It features dynamic sections that appear based on the user's selection, ensuring a smooth and intuitive experience.

You can check it by running it, and it has been deployed on Netlify. The link is provided below.

https://scintillating-sunflower-23198d.netlify.app/